requirements.txt
streamlit
pandas
python-docx
pypdf
google-generativeai
python-dotenv
